backend/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Authentication
djangorestframework-simplejwt==5.5.1
PyJWT==2.8.0

# Database
mysqlclient==2.2.0

# PDF Generation
WeasyPrint==60.1
Pillow==10.1.0

# Environment variables
python-decouple==3.8

# Date/Time utilities
python-dateutil==2.8.2

# Validation
jsonschema==4.20.0

# File handling
python-magic==0.4.27

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code quality
flake8==6.1.0
black==23.12.0
isort==5.13.2

# Production server
gunicorn==21.2.0

# Utilities
pytz==2023.3
python-slugify==8.0.1

setuptools